pyproject.toml
[project]
name = "frameit"
version = "0.1.0"
description = ""
authors = [
    {name = "Charles",email = "ch.husquin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "webcolors (>=24.11.1,<25.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]

[tool.poetry]
name='frameit'
packages = [{ include = "frameit", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]